Move src folder to page folder
