+++ pages/ErrorPage.module.scss
@@ -1,4 +1,4 @@
-@use "variables" as *;
+@use "../variables" as *;
 
 .ErrorPage {
   display: flex;
