+++ pages/Header.module.scss
@@ -1,4 +1,4 @@
-@use "variables" as *;
+@use "../variables" as *;
 .Header {
   width: 100%;
   position: sticky;
