+++ pages/Loading.js
@@ -1,4 +1,4 @@
-import classes from ./ styles / Loading.module.scss
+import classes from './Loading.module.scss'
 
 function Loading() {
   return (
