+++ pages/Loading.module.scss
@@ -1,4 +1,4 @@
-@use "variables" as *;
+@use "../variables" as *;
 
 .Loading {
   position: fixed;
