+++ pages/_app.js
@@ -1,0 +1,7 @@
+import "./App.module.scss"
+import "./index.css"
+function App({ Component }) {
+  return <Header lang={lang} setLang={setLang} />
+}
+
+export default App
