+++ pages/faq/Faq.module.scss
@@ -1,4 +1,4 @@
-@use "variables" as *;
+@use "../variables" as *;
 
 .Faq {
   width: 100%;
