+++ pages/faq/index.js
@@ -1,4 +1,4 @@
-import classes from './styles/Faq.module.scss'
+import classes from './Faq.module.scss'
 import { useTranslation } from 'react-i18next'
 import Accordion from '@material-ui/core/Accordion'
 import AccordionSummary from '@material-ui/core/AccordionSummary'
