+++ pages/i18nInit.js
