+++ pages/index.js
@@ -2,7 +2,7 @@ import { Route, Switch, HashRouter } from 'react-router-dom'
 import Home from './Home'
 import Releases from './Releases'
 import Faq from './Faq'
-import classes from './styles/App.module.scss'
+import classes from './App.module.scss'
 import Header from './Header'
 import { RecoilRoot } from 'recoil'
 import { useState } from 'react'
