+++ pages/releases/Releases.module.scss
@@ -1,4 +1,4 @@
-@use "variables" as *;
+@use "../variables" as *;
 
 .Releases {
   display: flex;
