+++ pages/releases/index.js
@@ -1,4 +1,4 @@
-import classes from './styles/Releases.module.scss'
+import classes from './Releases.module.scss'
 import { useTranslation } from 'react-i18next'
 
 function Releases() {
