--- src/States.js
@@ -1,3 +1,0 @@
-/**
- * Currently there are no states.
- */
